converted components check page into options page
