+++ mRemoteV1/UI/Forms/OptionsPages/ComponentsPage.resx
