+++ mRemoteV1/UI/Forms/OptionsPages/StartupExitPage.cs
@@ -36,7 +36,6 @@ namespace mRemoteNG.UI.Forms.OptionsPages
             Settings.Default.OpenConsFromLastSession = chkReconnectOnStart.Checked;
             Settings.Default.SingleInstance = chkSingleInstance.Checked;
             Settings.Default.StartMinimized = chkStartMinimized.Checked;
-            Settings.Default.StartupComponentsCheck = chkProperInstallationOfComponentsAtStartup.Checked;
         }
 
         private void StartupExitPage_Load(object sender, EventArgs e)
@@ -45,7 +44,6 @@ namespace mRemoteNG.UI.Forms.OptionsPages
             chkReconnectOnStart.Checked = Settings.Default.OpenConsFromLastSession;
             chkSingleInstance.Checked = Settings.Default.SingleInstance;
             chkStartMinimized.Checked = Settings.Default.StartMinimized;
-            chkProperInstallationOfComponentsAtStartup.Checked = Settings.Default.StartupComponentsCheck;
         }
     }
 }
