+++ mRemoteV1/UI/Forms/frmMain.cs
@@ -177,8 +177,6 @@ namespace mRemoteNG.UI.Forms
             Windows.TreeForm.Focus();
 
             PuttySessionsManager.Instance.StartWatcher();
-            if (Settings.Default.StartupComponentsCheck)
-                Windows.Show(WindowType.ComponentsCheck);
 
             Startup.Instance.CreateConnectionsProvider(messageCollector);
