+++ mRemoteV1/UI/WindowType.cs
@@ -1,4 +1,4 @@
-namespace mRemoteNG.UI
+﻿namespace mRemoteNG.UI
 {
     public enum WindowType
     {
@@ -16,6 +16,5 @@ namespace mRemoteNG.UI
         ExternalApps = 13,
         PortScan = 14,
         UltraVNCSC = 16,
-        ComponentsCheck = 17,
     }
 }
