+++ mRemoteV1/mRemoteV1.csproj
@@ -610,6 +610,9 @@
     <Compile Include="UI\Forms\OptionsPages\AppearancePage.cs">
       <SubType>UserControl</SubType>
     </Compile>
+    <Compile Include="UI\Forms\OptionsPages\ComponentsPage.Designer.cs">
+      <DependentUpon>ComponentsPage.cs</DependentUpon>
+    </Compile>
     <Compile Include="UI\Forms\OptionsPages\CredentialsPage.cs">
       <SubType>UserControl</SubType>
     </Compile>
@@ -815,8 +818,8 @@
     <Compile Include="UI\Window\BaseWindow.cs">
       <SubType>Form</SubType>
     </Compile>
-    <Compile Include="UI\Window\ComponentsCheckWindow.cs">
-      <SubType>Form</SubType>
+    <Compile Include="UI\Forms\OptionsPages\ComponentsPage.cs">
+      <SubType>UserControl</SubType>
     </Compile>
     <Compile Include="UI\Window\ConfigWindow.cs">
       <SubType>Form</SubType>
@@ -1111,8 +1114,8 @@
     <EmbeddedResource Include="UI\Window\BaseWindow.resx">
       <DependentUpon>BaseWindow.cs</DependentUpon>
     </EmbeddedResource>
-    <EmbeddedResource Include="UI\Window\ComponentsCheckWindow.resx">
-      <DependentUpon>ComponentsCheckWindow.cs</DependentUpon>
+    <EmbeddedResource Include="UI\Forms\OptionsPages\ComponentsPage.resx">
+      <DependentUpon>ComponentsPage.cs</DependentUpon>
       <SubType>Designer</SubType>
     </EmbeddedResource>
     <EmbeddedResource Include="UI\Window\ConfigWindow.resx">
